[code-infra] Move per-URL crawl work into worker threads
Move fetch, parse, link/target extraction, and HTML validation into a
single crawlWorker per URL. The main thread now only handles queue
management, deduplication, and post-crawl analysis.
